requirements.txt
# local package
# -e .

# external requirements
# click
# Sphinx
# coverage
# awscli
# flake8
# python-dotenv>=0.5.1

# DVC
dvc==3.33.3

pyyaml
pandas
scikit-learn